Move mininf test to advanced numeric
